Обработка сохранения пола и предпочтения

+++ src/main/java/ru/liga/tindertgbot/repository/UserFileRepository.java
@@ -5,7 +5,7 @@ import ru.liga.tindertgbot.entity.User;
 public interface UserFileRepository {
     void store(User user);
 
-    User load(int userId);
+    User load(long userId);
 
     void setStorageDirectory(String directory);
 }
